refactor: useCouponApplicabilityChecker selector로 구현 및 테스트 작성
