+++ src/recoil/coupon/selector.ts
@@ -1,8 +1,10 @@
 import { selector, selectorFamily } from 'recoil';
 
 import { couponsState } from './atom';
+import { orderTotalPriceState } from '../cartItems/selectors';
 
 import { Coupon } from '@/types/coupon';
+import { isCouponUsableTime, isCouponValid, isOverMinimumOrderAmount } from '@/utils/validations';
 
 export const couponListSelector = selector({
   key: 'couponListSelector',
@@ -21,5 +23,23 @@ export const couponSelector = selectorFamily<Coupon | undefined, string>({
       const coupons = get(couponListSelector);
 
       return coupons.find(({ code }) => couponCode === code);
+    },
+});
+
+export const applicableCouponSelector = selectorFamily<boolean, string>({
+  key: 'applicableCouponSelector',
+  get:
+    (couponCode) =>
+    ({ get }) => {
+      const coupon = get(couponSelector(couponCode));
+      const totalAmount = get(orderTotalPriceState);
+
+      if (!coupon || !isCouponValid(coupon)) return false;
+
+      if (coupon.minimumAmount && !isOverMinimumOrderAmount(coupon, totalAmount)) return false;
+
+      if (coupon.availableTime && !isCouponUsableTime(coupon)) return false;
+
+      return true;
     },
 });
